Split eval runner responsibilities
